% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{what_is}
\alias{what_is}
\title{Check if your number is nice or if it's trash- you've always wanted to know, and now you can!}
\usage{
what_is(x)
}
\arguments{
\item{x}{A numeric value between 0 - 2000. Don't go out of those parameters, or else!}
}
\value{
Returns a peer-reviewed, ABSOLUTELY final verdict on whether your number is nice or not.
}
\description{
Check if your number is nice or if it's trash- you've always wanted to know, and now you can!
}
\examples{
what_is(x)
}
